+++ src/library/DIPS.Mobile.UI/API/Builder/AppHostBuilderExtensions.cs
@@ -4,6 +4,7 @@ using DIPS.Mobile.UI.Components.Chips;
 using DIPS.Mobile.UI.Components.ContextMenus;
 using DIPS.Mobile.UI.Components.Images.Image;
 using DIPS.Mobile.UI.Components.Images.NativeIcon;
+using DIPS.Mobile.UI.Components.Lists;
 using DIPS.Mobile.UI.Components.Pickers.DateAndTimePicker;
 using DIPS.Mobile.UI.Components.Pickers.SegmentedControl;
 
@@ -21,6 +22,7 @@ using ContextMenuPlatformEffect = DIPS.Mobile.UI.Components.ContextMenus.Context
 using DatePickerHandler = DIPS.Mobile.UI.Components.Pickers.DatePicker.DatePickerHandler;
 using Image = DIPS.Mobile.UI.Components.Images.Image.Image;
 using ImageButton = DIPS.Mobile.UI.Components.Images.ImageButton;
+using ScrollView = DIPS.Mobile.UI.Components.Lists.ScrollView;
 using SearchBar = DIPS.Mobile.UI.Components.Searching.SearchBar;
 using SearchBarHandler = DIPS.Mobile.UI.Components.Searching.SearchBarHandler;
 using TimePickerHandler = DIPS.Mobile.UI.Components.Pickers.TimePicker.TimePickerHandler;
@@ -52,6 +54,7 @@ public static partial class AppHostBuilderExtensions
             handlers.AddHandler<Image, ImageHandler>();
             handlers.AddHandler<Button, ButtonHandler>();
             handlers.AddHandler<CollectionView, CollectionViewHandler>();
+            handlers.AddHandler<ScrollView, ScrollViewHandler>();
             
             AddPlatformHandlers(handlers);
         });