+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/iOS/CollectionViewHandler.cs
@@ -3,7 +3,7 @@ using UIKit;
 
 namespace DIPS.Mobile.UI.Components.Lists;
 
-public partial class CollectionViewHandler : Microsoft.Maui.Controls.Handlers.Items.CollectionViewHandler
+public partial class CollectionViewHandler
 {
     private static partial void MapShouldBounce(CollectionViewHandler handler,
         Microsoft.Maui.Controls.CollectionView virtualView)