+++ src/library/DIPS.Mobile.UI/Components/Lists/ScrollView/Android/ScrollViewHandler.cs
@@ -1,0 +1,15 @@
+using Android.Views;
+
+namespace DIPS.Mobile.UI.Components.Lists;
+
+public partial class ScrollViewHandler
+{
+    private static partial void MapShouldBounce(ScrollViewHandler handler,
+        Microsoft.Maui.Controls.ScrollView virtualView)
+    {
+        if (virtualView is ScrollView scrollView)
+        {
+            handler.PlatformView.OverScrollMode = scrollView.ShouldBounce ? OverScrollMode.Always : OverScrollMode.Never;
+        }
+    }
+}