+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## [16.3.0]
 - [SegmentedControl] Added segmented control.
 - [Label] Changed default text color to neutral_90.
+- [CollectionView] Added ShouldBounce property.
+- [ScrollView] Added ScrollView as a component, with ShouldBounce property.
 
 ## [16.2.2]
 - [Android][CollectionView] Fixed a bug where if you set ItemSpacing on a CollectionView, the CollectionView tries to snap to the elements.